+++ @ecomplus/storefront-template/CHANGELOG.md
@@ -3,6 +3,40 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.0.0-beta.106](https://github.com/ecomplus/storefront/compare/@ecomplus/storefront-template@2.0.0-beta.105...@ecomplus/storefront-template@2.0.0-beta.106) (2020-12-24)
+
+
+### Features
+
+* **cms:** preview for general settings page ([#367](https://github.com/ecomplus/storefront/issues/367)) ([d16409f](https://github.com/ecomplus/storefront/commit/d16409f4f7fea3cd5d300ca9804ac9e1931193f9))
+
+
+
+
+
+# [2.0.0-beta.105](https://github.com/ecomplus/storefront/compare/@ecomplus/storefront-template@2.0.0-beta.104...@ecomplus/storefront-template@2.0.0-beta.105) (2020-12-17)
+
+
+### Performance Improvements
+
+* **scripts:** use jsdelivr cdn for jquery also ([11920dc](https://github.com/ecomplus/storefront/commit/11920dcc5542e705c88bee91b285e30562d20678))
+
+
+
+
+
+# [2.0.0-beta.104](https://github.com/ecomplus/storefront/compare/@ecomplus/storefront-template@2.0.0-beta.103...@ecomplus/storefront-template@2.0.0-beta.104) (2020-12-16)
+
+
+### Features
+
+* **themes:** adding new niche flowers theme ([#374](https://github.com/ecomplus/storefront/issues/374)) ([07bbf0b](https://github.com/ecomplus/storefront/commit/07bbf0be6d2427177567c6deb7fd57039ca82fde))
+* **themes:** adding new niche flowers theme ([#375](https://github.com/ecomplus/storefront/issues/375)) ([9b5cef2](https://github.com/ecomplus/storefront/commit/9b5cef22da52d24132036940eb8c408a0d568511))
+
+
+
+
+
 # [2.0.0-beta.103](https://github.com/ecomplus/storefront/compare/@ecomplus/storefront-template@2.0.0-beta.102...@ecomplus/storefront-template@2.0.0-beta.103) (2020-12-16)